feat: move out traits to standalone component #79
