+++ src/features/Character/components/index.ts
@@ -1,0 +1,1 @@
+export { CharacterTraits } from './CharacterTraits/CharacterTraits';
